Deployment/requirements.txt
numpy==1.26.4
Pillow==10.4.0
streamlit==1.50.0
tensorflow==2.19.0
